--- IntrospectTests/AppKit/TestUtils.swift
@@ -1,17 +1,0 @@
-import Foundation
-import SwiftUI
-import AppKit
-
-enum TestUtils {
-    static func present<ViewType: View>(view: ViewType) {
-        
-        let window = NSWindow(
-            contentRect: NSRect(x: 0, y: 0, width: 480, height: 300),
-            styleMask: [.titled, .closable, .miniaturizable, .resizable, .fullSizeContentView],
-            backing: .buffered, defer: false)
-        window.center()
-        window.setFrameAutosaveName("Main Window")
-        window.contentView = NSHostingView(rootView: view)
-        window.makeKeyAndOrderFront(nil)
-    }
-}
