+++ IntrospectTests/AppKitTests.swift
@@ -1,7 +1,22 @@
+#if canImport(AppKit)
+
 import XCTest
 import SwiftUI
-
 @testable import Introspect
+
+enum TestUtils {
+    static func present<ViewType: View>(view: ViewType) {
+        
+        let window = NSWindow(
+            contentRect: NSRect(x: 0, y: 0, width: 480, height: 300),
+            styleMask: [.titled, .closable, .miniaturizable, .resizable, .fullSizeContentView],
+            backing: .buffered, defer: false)
+        window.center()
+        window.setFrameAutosaveName("Main Window")
+        window.contentView = NSHostingView(rootView: view)
+        window.makeKeyAndOrderFront(nil)
+    }
+}
 
 private struct ListTestView: View {
     
@@ -108,7 +123,7 @@ private struct SegmentedControlTestView: View {
     }
 }
 
-class IntrospectTests: XCTestCase {
+class AppKitTests: XCTestCase {
     
     func testList() {
         
@@ -184,3 +199,4 @@ class IntrospectTests: XCTestCase {
         wait(for: [expectation], timeout: 1)
     }
 }
+#endif
